Frontend Modularization: Refactoring into Pipelines into views, components and services.
