+++ frontend/src/modules/Pipelines/components/index.js
@@ -1,0 +1,6 @@
+export * from './PipelineCICD';
+export * from './PipelineEnvironmentCreateForm';
+export * from './PipelineEnvironmentEditForm';
+export * from './PipelineEnvironments';
+export * from './PipelineListItem';
+export * from './PipelineOverview';
