+++ frontend/src/modules/Pipelines/services/deleteDataPipeline.js
