+++ frontend/src/modules/Pipelines/services/deleteDataPipelineEnvironment.js
