+++ frontend/src/modules/Pipelines/services/getDataPipelineCredsLinux.js
