+++ frontend/src/modules/Pipelines/services/index.js
@@ -1,15 +1,9 @@
-export * from './browseDataPipelineRepository';
 export * from './createDataPipeline';
 export * from './createDataPipelineEnvironment';
 export * from './deleteDataPipeline';
 export * from './deleteDataPipelineEnvironment';
 export * from './getDataPipeline';
 export * from './getDataPipelineCredsLinux';
-export * from './getDataPipelineDag';
-export * from './getDataPipelineEnvironment';
-export * from './getDataPipelineFileContent';
-export * from './listDataPipelineBranches';
 export * from './listDataPipelines';
 export * from './updateDataPipeline';
 export * from './updateDataPipelineEnvironment';
-export * from './updatePipelineStack';
