--- frontend/src/services/graphql/DataPipeline/browseDataPipelineRepository.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const browseDataPipelineRepository = (input) => ({
-  variables: {
-    input
-  },
-  query: gql`
-    query BrowseDataPipelineRepository($input: DataPipelineBrowseInput!) {
-      browseDataPipelineRepository(input: $input)
-    }
-  `
-});
