--- frontend/src/services/graphql/DataPipeline/getDataPipelineDag.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const getDataPipelineDag = (DataPipelineUri) => ({
-  variables: {
-    DataPipelineUri
-  },
-  query: gql`
-    query GetDataPipelineDag($DataPipelineUri: String!) {
-      getDataPipelineDag(DataPipelineUri: $DataPipelineUri)
-    }
-  `
-});
