--- frontend/src/services/graphql/DataPipeline/getDataPipelineFileContent.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const getDataPipelineFileContent = (input) => ({
-  variables: {
-    input
-  },
-  query: gql`
-    query getDataPipelineFileContent($input: DataPipelineFileContentInput!) {
-      getDataPipelineFileContent(input: $input)
-    }
-  `
-});
