--- frontend/src/services/graphql/DataPipeline/listDataPipelineBranches.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const listDataPipelineBranches = (DataPipelineUri) => ({
-  variables: {
-    DataPipelineUri
-  },
-  query: gql`
-    query ListDataPipelineBranches($DataPipelineUri: String!) {
-      listDataPipelineBranches(DataPipelineUri: $DataPipelineUri)
-    }
-  `
-});
