--- frontend/src/services/graphql/DataPipeline/updatePipelineStack.js
@@ -1,10 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const updatePipelineStack = (DataPipelineUri) => ({
-  variables: { DataPipelineUri },
-  mutation: gql`
-    mutation updatePipelineStack($DataPipelineUri: String!) {
-      updatePipelineStack(DataPipelineUri: $DataPipelineUri)
-    }
-  `
-});
